refactor(cache): keep Cloudflare header policy in CDN adapter
