+++ packages/cloudflare/src/cache/cdn-adapter.runtime.ts
@@ -41,10 +41,36 @@ import type {
 } from "vinext/shims/cdn-cache";
 import type { CacheHandlerValue, IncrementalCacheValue } from "vinext/shims/cache";
 import { getRequestExecutionContext } from "vinext/shims/request-context";
-import {
-  clearCloudflareCdnResponseHeaders,
-  hasExplicitCloudflareNonCacheableResponsePolicy,
-} from "./cdn-response-headers.js";
+
+const NON_CACHEABLE_DIRECTIVE_RE = /\b(?:private|no-store|no-cache)\b/i;
+const CACHEABLE_EDGE_DIRECTIVE_RE = /(?:^|,)\s*(?:s-maxage|max-age)\s*=/i;
+const EDGE_POLICY_HEADERS = ["CDN-Cache-Control", "Cloudflare-CDN-Cache-Control"] as const;
+
+/** Remove every response header whose cache semantics are owned by Cloudflare. */
+function clearCloudflareCdnResponseHeaders(cacheControl: string): CdnResponseHeaders {
+  return {
+    "Cache-Control": cacheControl,
+    "CDN-Cache-Control": null,
+    "Cloudflare-CDN-Cache-Control": null,
+    "Cache-Tag": null,
+  };
+}
+
+/** Interpret Cloudflare's edge policy before core applies a replacement policy. */
+function hasExplicitCloudflareNonCacheableResponsePolicy(headers: Headers): boolean {
+  const edgePolicies = EDGE_POLICY_HEADERS.map((name) => headers.get(name));
+  if (edgePolicies.some((value) => value && NON_CACHEABLE_DIRECTIVE_RE.test(value))) {
+    return true;
+  }
+
+  const hasCacheableEdgePolicy = edgePolicies.some(
+    (value) => value && CACHEABLE_EDGE_DIRECTIVE_RE.test(value),
+  );
+  const browserPolicy = headers.get("Cache-Control");
+  return Boolean(
+    !hasCacheableEdgePolicy && browserPolicy && NON_CACHEABLE_DIRECTIVE_RE.test(browserPolicy),
+  );
+}
 
 /** The request-context cache surface this adapter relies on (narrowed from `unknown`). */
 type WorkersCacheLike = {
